constants: add tentative mappings for Ubuntu Touch 25.3.x
For now, use utnext images as branching doesn't yet happen. Once we
start branching in UT side, we'll update this mapping again to point
to image for that major version.

I *hope* that this verison will be final, and that we'll be able to
make a release by that time...