clickreviews: rename ubuntu-touch-{2505 => next}
Due to delays and ongoing discussions, we at the core team are still
unable to settle on what versioning scheme and what version number we
will use with the next version of Ubuntu Touch based on Noble.

To allow Clickable (and app developers) to side-step this issue, we've
introduced the concept of "next" framework, representing the rolling,
ongoing development of the next version of Ubuntu Touch. As indicated by
the full name "ubuntu-touch-next-internal", we provide absolutely no
guarantee on the stability of this framework. It however can be perfect
for testing the app against the upcoming release of Ubuntu Touch.

We apologize for any delays caused by this.

Bug: ubports/ubuntu-touch#2210